refactor: upgrades project to spring boot 3.4.1 latest
This commit included several changes to project's code and dependencies in order to upgrade it to
use the most recent Spring, Java and Jakarta EE technologies. For more details please refer to the
README.

BREAKING CHANGE: After upgrading to the latest Spring versions the app now requires JRE 17 to run.